Add functions to implement root mean prediction error
The rmpe function can also be implemented for distributional or spd matrix objects; keep in mind to use the correct metric in the context for the last step of computing the prediction error based on the test data and the estimate from the training data.

Do not use the functions estimate_ichimura() or estimate_2pred() functions if you want to just compute the rmpe. The new function rmpe_fn_ifr estimates the best direction based on the training data and computes the prediction error based on the test data; along the way bw and M are chosen adaptively. This new function already implements the same method as the function estimate_ichimura does.